pyoxidizer: build managed Rust projects with --locked
As of the previous commit we should write out a Cargo.lock that
is current and doesn't need any changes in order to build. Therefore
it should be safe to run with --locked so we get deterministic crate
versions matching those from the Cargo.lock in this Git repository
that PyOxidizer was built from.

This is a partial solution for #364.