pyoxidizer: add --env and --env-var arguments for injecting variables into Starlark
This feature has come up a few times as something that would be useful
to have. I know there are GitHub issues related to it. But I can't find
a single one tracking the feature, so we don't have any issue
annotation.

Anyway, the feature is pretty self-explanatory and hopefully
self-documenting via the command help. This should definitely make
it easier to make configuration files more powerful based off of
external inputs.