tugger: define a global context
We've done this for build targets and PyOxidizer to facilitate the
use of global state inside Starlark since the current version of the
crate doesn't support global variable access. This is hacky. But it
appears the only way given current Starlark APIs.

Anyway, this will be used to register our code signers globally
within the environment.