pyoxidizer: change how variable injection works
Injecting top-level variables doesn't work very well because you
can't access undefined symbols and there's no way to test if a symbol
is defined.

There was also a bug where the CLI arguments didn't actually work
because we were inspecting the wrong ArgMatches instance. This one
is hard to test and that's on me for not verifying the functionality
works as intended.

I'm glad I caught this before releasing. I should have tested this
better before pushing the feature.