cryptographic-message-syntax: sort certificates in SignedData so issuer comes first
I'm not sure if this is generally a best practice or not. But it is the
order appearing in Apple's CMS messages. I don't think it matters
to most readers, as certificates get loaded into a bag of certs. But
it will help compare the output of our CMS messages with Apple's.