tugger-apple-codesign: limit <cdhash> value to 20 bytes of digest
It appears the digest in the XML is always the first 20 bytes of
the digest. However, stronger digests can be specified via a separate
mechanism. This will be implemented in a separate commit.

In addition, we were improperly encoding the value as <string> instead
of <data>. This has also been fixed.