tugger-apple-codesign: include SHA-256 code directory signed attribute
I ran into problems validating bundles signed with this code. Comparing
the CMS signatures revealed that codesign was including an additionl
signed attribute with an OID + octet string. The OID indicated it was
a SHA-256 digest. And experimentation revealed that the octet string
was the SHA-256 digest of the code directory.

So this commit implements support for writing that missing attribute.