tugger-apple-codesign: embed Apple certificate data
We've written code for finding Apple certificates. I think this is
worth more trouble than it is worth. Let's just embed all the
Apple certificates so we can readily access their data.

As a plus, this gives us the opportunity to define our own
AppleCertificate trait to provide Apple-specific extensions to
parsed X.509 certificates. I have plans for this in upcoming
commits. But we start by embedding the certificates and defining
a skeleton trait.