tugger-apple-codesign: define a bunch of missing OIDs and enumerations
We want to shore up support for handling code signing certificates.
As part of that we define a bunch of missing OIDs governing how
these certificates work and add enums to refer to them. There's
still a lot of room to build out functionality. But having this
data defined is better than not having it.