x509-certificate: move certificate type out of cms crate and refactor
This commit extracts the high-level certificate handling code from the
CMS crate and significantly refactors it.

We now provide an X509Certificate that is a thin wrapper around a
tracked ASN.1 data structure.

We extend that through CapturedX509Certificate, which holds a reference
to the data it was parsed from. Having access to the original data
enables us to validate signatures with confidence.