x509-certificate: shore up algorithm handling
There were a handful of subtle bugs in our definition of algorithms.

We were generally ignoring parameters on AlgorithmIdentifier and
making assumptions about what sub-algorithm (e.g. curves in the case
of ECDSA) was being used. We also made assumptions that the
verification algorithm always mapped to certain key algorithms.

This commit hopefully teaches the code to be much more resilient
about the relationship and existence of various key, signing, and
verification algorithms.

The ECDS key algorithm enum now captures the elliptic curve it is
associated with. We now properly serialize the curve info to
AlgorithmIdentifier and parse it back when reading certificates.

Similarly, when we sign data with a key pair, we return an instance
of SignatureAlgorithm that exactly defines what signature ring
actually used, rather than leaving it up to the caller to guess
or make assumptions.

When obtaining a concrete ring verifier algorithm, we require the caller
to pass in both the key algorithm and the signature algorithm, as
only certain combinations are logical.

The new code should be much more resilient than before. Tests have been
added to verify that different ECDSA curves work as expected and that
we are able to verify signatures where the signing key's curve "bitness"
doesn't match that of the signature. (ring doesn't support creating
these signatures but it supports verifying them.)