tugger-apple-codesign: automatically attempt to derive designated requirements
This commit adds a function for resolving a designated requirement
expression from a given code signing certificate. It attempts to map
the different types of official Apple code signing certificates to
requirement expressions in a way that matches what codesign does.

We also change the settings code so automatic requirements derivation
is enabled by default.