tugger-apple-codesign: attempt to automatically register Apple CA certs
This seems like the correct things to do by default, since binaries
will fail trust verification without the certs present.

With this change, Apple binary and bundle signing appears to "just
work." However, only if timestamp servers are not in use. Apple's
CMS parser has proved to be quite finicky and I suspect there's
some nuanced difference with that part of the CMS payload. We'll
have to address this in upcoming commits.