x509-certificate: rework AlgorithmParameter handling
This commit starts by reverting ec985e5bf1239fb32449330cf522496430b531fe.

I think the use of Option<T> here for parameter handling is
cleaner, as it doesn't require callers to pass in an encoding
mode, which can be different depending on the context.

Rather, we rework the implementation of Values so it constructs
a new Constructed with the NULL tag if None is stored. The old
encode_ref() function is gone and encoders can simply pass a
reference to the instance, as it implements Values.