python-packaging: normalize package distribution directory names
The on-disk directory for Python package distributions has the
package component of the name normalized to lowercase with hyphens
replaces with underscores. See importlib.metadata.Prepared.__init__
for the canonical code in the Python standard library.

This commit teaches our filename resolution code to apply the
same normalization.

Tests have been added for the filename resolution and that the
filesystem scanner preserves the proper package name (which was
already the case because it keys off the parsed metadata files
instead of the directory name).

Closes #394.