cryptographic-message-syntax: manually implement Debug
Debug printing of certificates and CMS data structures was difficult
to impossible to read because output spanned hundreds or thousands
of lines due to Vec<u8> printing. And there were quality of life
issues, such as OIDs being printed as byte strings.

This commit implements Debug on various types so the printed
structure is much cleaner. Notably, byte strings are printed as
hex encoded strings and OIDs as their X.Y.Z string form. Debug
output is actually readable now. And it is relatively easy to
copy values and play around with them.