cryptographic-message-syntax: properly compute content that is signed
Before, we were incorrectly determining the content that was to
be cryptographically signed when there was encapsulated content *and*
signed attributes. Just signed attributes worked. But not the
combination of them.

RFC 5652 says that the message digest is computed over the encapsulated
content or the signed attributes, not both.

The incorrect computation of what content was to be signed was leading
to failure to verify time-stamp token SignedData, which are well-formed.

This commit also introduces APIs for obtaining the time-stamp token
SignedData and adds tests that verify its verification.

As part of going down this rabbit hole, I discovered that there is
yet another bug lingering in verification land. Possibly due to choosing
the wrong algorithm in places. This is causing the vendored Firefox
Apple code signature CMS blob to fail time-stamp token verification.
Although the bug could be Apple wonkiness: we'll have to poke around
some more.