tugger-apple-codesign: start to implement trait for interfacing with Mach-O
I want to refactor the low-level Mach-O parsing and manipulation code
so it lives in a single place and is more flexible. Rather than
define separate types, I'm going to define a trait that extends
the functionality directly on MachO instances.

We start with calculating the executable segment boundaries, as
that is a simple, standalone function today.