tugger-apple-codesign: support writing new signatures
Before, we required a binary to have an existing signature for us
to write out a signature. After this commit, we can write signatures
on binaries that didn't have them before.

It turns out this was simpler than I thought it would be.

The reason I avoided this problem initially was because I thought
writing a new load command would entail updating offsets throughout
the Mach-O. Turns out that Mach-O binaries typically contain a lot
of NULL bytes between the load commands and the next meaningful
data (the first section within a segment). So in many cases we can
just write new load commands without having to worry about
accounting.

For reasons unclear to me, we had to increase the amount of padding
bytes in the signature otherwise we didn't have enough room to write
the full signature. I'd like to get to the bottom of this, but that's
for a separate commit.