tugger-apple: move universal Mach-O creation code to crate
We want to expose universal Mach-O creation to Tugger via Starlark.
Let's move the code out of the code signing crate to somewhere more
generic.

As part of this, we also rewrite the logic for writing out the binary.
We now create the fat header data without consulting the original
version.