tugger-wix: consistently compute component IDs
Before this change, we constructed a new PathBuf to compute the relative
path. This wasn't guaranteed to yield the same path as what was stored
in the FileManifest and could result in divergence of the derived
IDs/GUIDs.

After this change, we should consistently use the path stored in the
FileManifest for ID derivation.