tugger-wix: use full file relative path for deriving File Id
Before, we were only using the filename part of the path to derive
the <File Id> and GUID values. If there were multiple files with the
same filename, there would be ID/GUID collisions.

This commit teaches the derivation to use the full relative path
of the file in the FileManifest as the value to derive against.