release: force use of path reference to for pyembed during release
If we don't do this and we are releasing the pyembed crate, the
Cargo.toml attempts to reference a pyembed version on the registry that
hasn't been published yet.

After this change, we force the use of a local path reference and then
we update to use a registry reference after it is available.