pyoxidizer: set RUSTC in build environment
The output in #380 seemingly indicates that cargo can't find `rustc`.
I hypothesize this is because `rustc` isn't in PATH.

Cargo listens for the RUSTC environment variable to define the Rust
compiler to use. Since we now resolve the paths to cargo and rustc,
I think it makes sense to force cargo to use the path to rustc that
we resolved.

This may not completely fix builds not working. If it does, great. If
not, we'll need to go a step further and adjust PATH so our Rust
toolchain with cargo/rustc is the first item in it.