pyoxidizer: target appropriate Apple SDK when compiling custom config.c
Before this commit, our invocation of the cc crate would likely pick up
the default Apple SDK. This would work most of the time. However, during
scenarios like cross-compiling, it may not work. For example, if the
default SDK is a 10.15 SDK and we are compiling for aarch64, which is
only supported by the 11.0 SDK. In this case, you'd get a compiler
error about the supported target not being supported. Even if things
did work, building of the final binary would trigger full Apple SDK
checks and this would reveal an error.

This commit teaches the config.c compiling to use the same logic as
the Rust project / final binary building code for finding and validating
the Apple SDK. It should enable errors to be detected sooner. It
should also hopefully prevent hard-to-debug compiler errors from
occurring, instead presenting an actionable error about the Apple
SDK.