pyoxidizer: simplify mtime adjustment on archive extract
By default, the tar extraction will normalize mtime to the stored
value. We added supporting for adjusting mtimes post extract a few
commits ago. Thinking it over, just not doing any mtime adjustment
is the simplest solution, as the mtime of the created file will be
"now" by default and this is sufficient to mitigate the old mtime in
archives problem.