tugger: resolve relative paths when writing wheel files
Some Starlark functions accept a path argument. Without any
normalization, relative paths will be relative to the process's
current working directory. We have a convention of wanting relative
paths to be relative to the configured build path.

This commit introduces a function on the context to normalize
paths. Absolute paths are preserved. Relative paths are joined with
the current build path.