pyoxidizer: add and use helper for resolving default target triple
We've been using HOST in most places. But HOST represents the build
environment, not the runtime/target environment. TARGET is a more
appropriate triple to key off of.

As part of this, we ensure we still use GNU Linux from musl libc
binaries, as that is backwards compatible and more user friendly.

All use of HOST has effectively been replaced with TARGET. And we
no longer make HOST available via env!.