pyoxidizer: use Arc<Mutex<T>> wrappers for Python resource types
This required extensive refactoring of the trait for managing
the resource context, as we now need to hold that type in a wrapper
type in a mutex.

I probably could have split this out into multiple commits. But the work
got away from me and I ended up doing everything in one go.