pyembed: support calling multiprocessing.set_start_method() automatically
pyembed is taught to call multiprocessing.set_start_method()
automatically when servicing the import of the multiprocessing module.

Support for this configurable behavior is wired through to Starlark.

On macOS, we use the "fork" method (unlike Python itself) because
we don't use framework Python builds and aren't subject to the
limitation of fork() not working in framework code.

multiprocessing still doesn't "just work" on Windows. That will
require additional feature work.

Part of #111.