pyembed: support --multiprocessing-fork dispatch
When multiprocessing uses the "spawn" method to start processes,
it invokes a new process with arguments --multiprocessing-fork [args].

This commit teaches the main dispatch code to look for the presence
of this sentinel argument and to call into the multiprocessing module
accordingly.

This change effectively makes multiprocessing "just work" on Windows,
without applications needing to do anything special to handle
spawned processes.

The new behavior is configurable via Starlark. And documentation has
been updated to reflect caveats.

Closes #111.
Closes #362.