apple-xar: add project skeleton
In order to read and write Apple flat packages, we'll need to read and
write XAR archives. This commit adds a project skeleton for a generic
crate to handle XAR archives.

There is an existing `xar` crate but it is incomplete and not heavily
used. It will be easier to maintain our own crate.