cpio-archive: add skeleton for new crate
In order to read Apple flat packages, we need to implement support for
reading cpio archives.

Unfortunately, I couldn't find an existing cpio crate that satisfies our
needs. The only crate I found was `cpio` and it is 3+ years old and only
implements one of the cpio archive formats. Unfortunately, this isn't
the cpio archive format that is always used by Apple flat packages. So,
we'll have to roll our own crate.