tugger-file-manifest: support more coercions into FileEntry
I want to add more functionality to file entries to make them
more powerful. e.g. the ability to store symlinks.

Before this commit, we had a ton of call sites in this repo that
constructed instances directly. This meant that changing the fields
was an expensive refactor.

This commit adds a bunch of new coercions to FileEntry. We also
introduce constructor functions to instantiate instances.

Various call sites have been updated to use the new functionality,
greatly reducing the dependence on the struct's fields.