tugger: allow overriding URLs for e.g. WiX Toolset via environment variables
I need to run PyOxidizer in an environment without access to the
internet (for security reasons). That fails when PyOxidizer tries to
download WiX Toolset. This patch adds a way of overriding the
hard-coded URL via environment variables. For example, I will set
`WIX_TOOLSET_URL=<some path>` in my build environment.

This is related to issue #204.