tugger-wix: store arch instead of target triple on WiXInstallerBuilder
This is more specific about what the value is being used for.

In order to prevent the Starlark constructor from raising outside
of supported WiX targets, we default to `x64` on supported
target triples. The logic of which target/arch to use needs
an overhaul. But this will be for subsequent commits.