pyembed: port to PyO3
This started out as a series of commits. But about 3-4 commits in
the hackery required to get cpython and pyo3 to play well together
was too much and writing microcommits proved to be too difficult.
So I just squased everything together into one giant commit.

This commit will break PyOxidizer in some configurations because the
crate configuration mechanism changed. This will be addressed in
a subsequent commit: this commit is already too large as it is.