pyembed: ensure GILGuard is dropped before interpreter is finalized
There is a subtle race condition between finalizing the interpreter
and dropping the GILGuard. This doesn't materialize with the cpython
crate because its Drop doesn't do anything special. But it matters to
pyo3, which attempts to decref pointers in its GIL "registry." If
it attempts to do so after the interpreter is finalized, we likely
get a UAF and segfault.