pyoxidizer: rewrite build logic to use PyO3
This commit finishes the port of PyOxidizer to PyO3.

This entails a lot of changes. The gist of it is we write out
a PyO3 configuration file to configure that crate. Then we retool
pyembed and various Rust project building code to use that config
file and PyO3's configuration mechanisms. There are a lot of
changes in here.