pyoxidizer: remove out_dir from link library search path
The only caller of this function (`resolve_python_linking_info()`)
passes in a temporary directory for `out_dir` and reads the
content of the built files immediately after this function is
called. The temporary directory is long deleted by the time the
cargo metadata lines are used.

So I think removing this line should be harmless, as it wasn't
actually accomplishing anything.