pyoxidizer: return library data from link_libpython()
The only caller of this function passed in a temporary directory
for the output directory and read the file data immediately after
calling the function. So returning a PathBuf didn't make much sense.

This commit refactors things to return the raw library data and
its filename, simplifying the code a bit.