pyoxidizer: create temp directory in link_libpython()
The previous commit changed the API so we return the library
data directly instead of the path to it. This means we no
longer need to pass an output directory into the function and
we can use a temporary directory as an implementation detail.

We were already using a temporary directory in this function. In order
to ensure things didn't conflict, we use 2 sub-directories for the
distinct operations.