pyoxidizer: remove linking annotation to self from link_libpython()
We were emitting an annotation that effectively said to link against
ourselves. Because the library data isn't manifested as a filename
any more, I think it makes sense to remove this annotation and have
it added later, when we actually want to link against it.

As part of this, we move the annotation to the first, as I think that
makes sense.