pyoxidizer: centralize the definition and rationale of pythonXY name
There is some nuance as to why things are the way they are. We had
a comment for this in libpython.rs. But that isn't the most appropriate
place for it.

We also had `pythonXY` repeated a few places. Defining a new accessor
function seems like the appropriate place to hang the documentation
off of.