pyoxidizer: emit library search path as part of link settings
As part of refactoring this code, I noticed some code in
project_building.rs that was using RUSTFLAGS to adjust the linker
search path. This seemed wrong to me, as Python linking annotations
should all be handled via `cargo:` lines.

This commit unifies things to emit the annotation for the dynamic
library search path in the same place we emit the annotation for
the static library search path.