pyoxidizer: use an enum and trait to express linking variations
Previously, we had a single struct hold a union of states requires
to express the libpython linking configuration. The linking
configuration can be nuanced and only very specific states are
supported.

This commit refactors the expression of our link settings to be
captured by dedicated structs with an enum + trait to make behavior
generic. This results in a bit more code. But I think it makes the
semantic intent more clear.