pyoxidizer: stop writing cargo_metadata.txt file
This file existed primarily to emit linker annotations needed to
supplement the libpython linking since the cpython/python3-sys crates
didn't fully configure linking in the context of PyOxidizer since
we produced a custom libpython.

After the switch to PyO3, the PyO3 config file generated by PyOxidizer
contains a full linking description and cargo_metadata.txt is no longer
needed for linking.

The only thing this file was used for was to annotate the location of
the autogenerated Rust file containing the interpreter config. This
is trivial to inline in places that need it. So we do that. This
makes a lot of code go away.