oxidized-importer: extract importer code to own crate
Previously, the pyembed crate defined both the oxidized_importer
extension and types and our interpreter embedding API. The
oxidized-importer crate contained a hacky re-export of just the
extension bits of pyembed.

The complexity of pyembed grew over time and we eventually
wanted to make oxidized_importer a standalone entity. But it
was still part of pyembed.

Now that the switch to PyO3 enabled pyembed to be a normal
Rust crate, it frees us up to implement crates in a more
traditional manner.

This commit extracts the extension parts of pyembed to the
oxidized-importer crate. But because we now need to publish this
crate on crates.io, we rename the crate to python-oxidized-importer,
as that name makes more sense in the context of crates.io.