python-packaging: implement support for writing zip applications
Python has a built-in mechanism for executing applications as zip files.
This commit implements support for writing zip files defining Python
applications. It is essentially a reimplementation of Python's zipapp
module.

A key difference is our implementation is more deterministic, as
zip archive members are always sorted and the mtime can be defined
explicitly.

Part of #429.