release: disable pyembed default features when generating lockfile
I'm going to introduce a feature that pulls in additional crate
dependencies. This will likely complicate our Cargo.lock story in
the future. But for now, we'll keep the feature disabled in PyOxidizer.

This commit effectively makes the lock file build configuration
match how PyOxidizer builds the Rust crate. So when we introduce
optional dependencies in the pyembed crate, the lock file shouldn't
pick them up.