pyoxidizer: log the actual resource collector behavior
See the changelog entry. Before, log messages were deceptive and
could be wrong. After, the collector returns what actions it
performed and we report these actions to the end-user.

There is infinite potential to improve the level of detail
here. But we start simple by recording basic resource operations
such as ignore and add.