pyoxidizer: use same major-minor version for host distribution in tests
Previously, we used the default host distribution when iterating through
all distributions in some tests. This resulted in a mismatch for some
combinations. For example, we used a host Python 3.9 when testing
against a target 3.8. This behavior could lead to incorrect results
in some tests. For example, compiling source to bytecode would use
the wrong Python version.

This was all fine for existing tests. However, the introduction of
Python 3.10 results in existing tests failing because they attempt
to use Python 3.9 to compile a source file with unrecognized 3.10
syntax.

I don't believe this issue exists outside of tests because Starlark
resolves a host distribution with a matching major minor version.