docs: restructure top-level project docs
The repository has grown to a collection of projects but the top-level
doc structure was still PyOxidizer centric. This commit changes it
so the top-level docs highlights the varied projects within and
presents things as largely separate projects.

While we're here, we also add tags to each Sphinx config file so we
can potentially differentiate docs based on what is built.