pyoxidizer: release 0.19.0
This could arguably be 0.18.1. But I don't feel like going through
and changing a bunch of crate release numbers to do the release. We
don't really abide by semantic versioning anyway and a minor
version release incurs about the same amount of work as a major
one. So why not call 0.19.0?