tugger-debian: implement PGP cleartext signature verification
Debian packaging uses RFC 4880 Section 7 "Cleatext Signature Framework"
signatures for PGP signatures. The pgp crate doesn't (yet) have this
functionality implemented. (This is tracked by
https://github.com/rpgp/rpgp/issues/122.)

This commit implements RFC 4880 Section 7 signature verification.

The bulk of the new code is in a new pgp module. The code was written
in an isolated manner to make it easier to port to the pgp crate (if
there is interest in doing so).

There are a few sub-optimal aspects to this code:

* Hasher handling. The digest being verified is a concatenation of the
  cleartext and signature metadata. I didn't want to re-hash the full
  cleartext for each signature for performance and simplicitly reasons.
  However, the pgp crate doesn't implement Clone on its hashers. So
  I had to reinvent the hasher wheel to enable this code to clone()
  hasher state.
* Signature reading. We currently buffer the entire signatures data
  from the reader. It is certainly possible to stream this data in.
  But for now the simple solution wins.

I haven't thoroughly tested this code. No doubt there are edge cases.
But as the added tests show, it does appear to verify a signature
from a Debian InRelease file and reject the signature if the file is
tampered, so I'm pretty confident it is working in some scenarios.