tugger-debian: implement more granular control file parser
Previously, we accumulated a Vec<ControlParagrah> from a reader.
This is inefficient when reading large sources.

This commit refactors the parsing logic to work on a single
paragraph at a time. The existing reader is implemented in terms
of the new code.

The new parser isn't super robust and there are some edge cases.
But this is a step in a desired direction.