tugger-debian: implement constraint checking; fix arch parsing bug
Now that we have support for parsing version strings, it is possible to
implement constraint checking against parsed dependency expressions.
This commit lays the groundwork for performing these checks.

As part of writing the tests, I discovered that architecture syntax
parsing wasn't working. So we fixed that and added test coverage.