tugger-debian: honor `Acquire-By-Hash: yes` setting
Modern Debian repositories set an `Acquire-By-Hash: yes` field
to trigger clients to fetch URLs containing the content digest of the
entity they are fetching. We had support for parsing this field but
weren't keying off of it.

This commit implements functionality for resolving the `by-hash` path
for a release file entry and automatically using it if the release
file says to use acquire-by-hash mode.

Pre-existing tests should now fetch teh by-hash URLs. The tests
pass, indicating this functionality apparently works.