tugger-debian: add higher-level API for control.tar file reading
Previously, we just exposed control.tar files inside .deb packages as
vanilla tar archives. This commit introduces some new types to more
strongly define the content of control.tar files.

It doesn't do much beyond classifying files as an enum and parsing
the control file. But it is a start.