pyembed: properly pass `pipe_handle` argument to `multiprocessing.spawn.spawn_main()`
We were constructing a value to use as the argument but weren't actually
passing that argument. Oops.

We lack test coverage for this spawn method because of wonkiness with
the Rust-based test runner. It is probably possibly to implement an
integration test for this. We'll leave that for another time.

Fixes #483.