tugger-debian: refactor Packages retrieval into a Trait
The mechanism for retrieving files from parsed Release files should
be transport agnostic. We should therefore be able to define the
functionality in traits and have concrete implementations simply bring
the I/O in order to implement.

This commit moves the logic for retrieving Packages files to a new
`ReleaseReader` trait to get us closer to that vision.