debian-packaging: remove ReleaseFileDigest type
It was largely redundant with the newer and more generic ContentDigest.
As part of this we parse the hex value when resolving the entry. This
may add some overhead when parsing indices. But it catches errors sooner
and enables us to not have to carry a 2nd digest type around.