debian-repo-tool: implement repository-copy command
I haven't tested this very thoroughly yet. But it seems to work under
limited testing.

This required teaching debian-packaging to resolve strings into
repository readers and writers. And we introduced support for defining
a configuration struct for repository copying, which we use serde
to easy read from a config file.