global: upgrade zstd crate to 0.9
This was downgraded to 0.7 several days ago to provide compatibility
with async-compression, which was using the old version. A new version
of that crate was released today using a newer zstd crate version.
So we upgrade back to 0.9.

We again target the canonical rpm-rs crate since we don't need to
carry a zstd downgrade patch. And this means linux-package-analyzer
is releasable to crates.io, so we add it to our release script.