pyembed: remove acquire_gil() and release_gil() from MainPythonInterpreter
Recent commits introduced and moved to a `with_gil(|py| ...)` function
for running a function with the GIL held. This is now the preferred
way to run code with the GIL held. It is also the direction the PyO3
API is likely headed.

With no consumers of the old `acquire_gil()` API remaining, this
commit removes that API and storage of the GILGuard.