pyembed: release the GIL when initializing interpreter
Before, `MainPythonInterpreter` held onto the GIL post interpreter
initialization. The GIL was effectively held for the lifetime of
instances and never releases.

This behavior was likely introduced from the very first version of
this code. I don't think it was intentional and any documentation of
its behavior was likely only a side-effect of what was going on
since I can't recall actively thinking about whether the GIL should
be held or not!

With previous commits shoring up the API around calling into the
interpreter and not having `MainPythonInterpreter` hold a GIL
guard, I think it makes sense for the GIL to not be held by
`MainPythonInterpreter`. This is ultimately a more flexible API,
as callers can move `MainPythonInterpreter` to another thread
without worrying about the GIL.

As part of this, we also found some corner cases where we weren't
asserting the GIL was acquired and/or the interpreter wasn't
finalized. These cases were flushed out by existing tests once we
started releasing the GIL.

Closes #500.