oxidized-importer: properly handle borrowed references during multi-phase init
Single-phase and multi-phase initialization don't agree on whether
the module creation function returns a new or borrowed reference.
Previously, our code assumed the existence of a new reference. When
multi-phase initialization was being used, we would misinterpret a
borrowed reference as a new reference and cause refcounts to get
out of sync. This could lead to a use-after-free.

We refactor the code around module creation to account for this
difference. Because behavior is very subtle, verbose comments
have been added.

Closes #490.