pyoxidizer: set SETUPTOOLS_USE_DISTUTILS=stdlib when modifying distutils
Without this, setuptools may use its own vendored distutils instead of
our copy. This could result in build failures.

There might be some bug reports on the issue tracker related to this.
But I found this when attempting to adopt the latest Python standalone
distributions. These pull in the latest pip/setuptools, which appear
to be more aggressive about using a vendored distutils.