pyembed: upgrade pyo3 to 0.16
A new version is out. Let's stay modern.

We have to upgrade all references throughout the workspace since this
crate needs to be in sync everywhere.

The new version dropped the `resolve-config` crate feature by default.
This had fallout in our `Cargo.toml` files.

The `WITH_THREADS` config option was also dropped, since upstream
dropped Python 3.6 support.

`PartialEq` was also dropped and replaced with `.is()`, `.eq()`, and
a handful of other functions. We update our code accordingly.

I _think_ some of our use of `==` meaning pointer equality testing was
subtly wrong. So this has been updated as part of this change.

There are a handful of deprecation warnings being emitted during builds
now. We'll fix that in future commits.