apple-codesign: functionality for resolving a binary's targeting settings
We want to enable code signing to apply better code signing settings
based on the minimum OS version being targeted. In order to make this
automatic, we'll want to sniff out targeting settings of a binary.

This commit implements a function for reading the mach-o loader commands
which indicate version requirements and turn them into a Rust data
structure. We don't yet do anything meaningful with the data except
expose it through an `rcode extract` data view.