apple-codesign: use mach-o binary target settings to influence CD version
One of the issues reported in #513 is that iOS 15+ requires a modern
code directory version.

Building upon the functionality implemented in the previous commit,
this commit feeds the binary targeting info into the `adjust_version()`
function and will automatically increase the minimum version for
binaries targeting modern operating systems.

The logic for when to clamp is likely incomplete and will need refined
over time. But for now it seems like a good start.