apple-codesign: store entitlements as parsed plist
We're slowly working our way towards DER encoded entitlements.
In order to get there, we want to internally store entitlements
as a parsed plist so we can easily encode them to DER.

This commit switches storage of entitlements from a raw string
to a parsed plist. In doing so, we normalize the entitlements
plist XML in the signature. I think this should be fine. But it
is a change in behavior.