x509-certificate: add a convenience API for verifying with a digest algorithm
This came up in the context of investigating #528. Ultimately, the
verification error was due to the reproduce code getting verification
subtly wrong. Crypto is subtly complex!

This commit introduces a new API on our X.509 certificate type.
We also introduce tests on the CMS crate with the reproduce case
so we have coverage over a mismatch between the key's digest algorithm
and the actual digest algorithm.