cryptographic-message-syntax: add SignerInfo method for signature verification with content
This commit improves the ergonomics of performing cryptographic
signature verification against content defined outside the CMS
payload. As #528 shows, this is apparently something that happens
in the wild. Hopefully the existence of a separate API will enable
people to find and use this method instead of reaching for lower-level
x.509 certificate verification, which may ignore attributes on the
CMS signature indicating which verification algorithm to use.

Closes #528.