pyoxy: use libtest-mimic in custom test binaries
By implementing a proper test harness in the test binaries we get
compatibility with nextest. See also
https://github.com/nextest-rs/nextest/issues/122 for why we need to do
this. This is probably a good idea in the absence of nextest anyway.

Note that libtest-mimic doesn't appear to honor `--ignored` in `--list`
mode. This might be a bug in their harness. I'm unsure.